Rename msg file to d_graph_update
